Partner Portal: Add the 'credit-card' redux store
This PR adds the base store for the 'credit-card' payment method store that is going to be used in the Partner Portal. It's pretty much a port from what's currently used in the composite-checkout package and might need improvements later.

This is going to be used behind a feature flag, so it won't harm having this merged