Fix stubbing for `is_automattician`
The previous implementation created a chicken-and-egg situation with the test env and broke the logic in production.

In the ETK PHP test env, `is_automattician` is not available globally. At first, I defined it under the `A8C\FSE\ErrorReporting` namespace in the test. This worked for the stubbing purposes, but required me to change the `function_exists` call in the implementation to include the namespace in order for it to work as expected in the test env. I thought that in production, this check would work, too, but it didn't. The function is defined globally in production and the check for it under that namespace failed, meaning that the `should_activate_sentry` always returned `false` in production.

The solution was to remove that `function_exists` check which was creating more trouble than it's worth, and make sure the function was available in the test env, for all tests.